kiemtra01-codex-TieuLuan01/services/user_service/requirements.txt
Django>=5.2,<5.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
requests==2.33.1
PyMySQL==1.1.2
cryptography==46.0.2
gunicorn==23.0.0
